requirements.txt
pyautogui
requests
schedule
python-dotenv
pillow
pyscreeze
telebot